Добавил базовый класс для презентеров.
