+++ app/src/main/java/com/volokhinaleksey/dictionaryofwords/repository/ApiHolder.kt
@@ -3,9 +3,7 @@ package com.volokhinaleksey.dictionaryofwords.repository
 import com.google.gson.FieldNamingPolicy
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
@@ -32,23 +30,22 @@ object DictionaryApiHolder : ApiHolder {
             GsonBuilder().setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
                 .create()
 
-        val okHttpClient = OkHttpClient.Builder()
-        okHttpClient.addInterceptor(BaseInterceptor())
-        okHttpClient.addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-
         Retrofit.Builder()
             .baseUrl(BASE_URL)
             .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
             .addConverterFactory(GsonConverterFactory.create(gson))
-            .client(okHttpClient.build())
+            .client(createOkHttpClient())
             .build()
             .create(ApiService::class.java)
     }
 
-}
+    private fun createOkHttpClient(): OkHttpClient =
+        OkHttpClient.Builder()
+            .addInterceptor(
+                HttpLoggingInterceptor().setLevel(
+                    HttpLoggingInterceptor.Level.BODY
+                )
+            )
+            .build()
 
-class BaseInterceptor : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        return chain.proceed(chain.request())
-    }
 }
