+++ app/src/main/java/com/volokhinaleksey/dictionaryofwords/ui/dictionarywords/DictionaryOfWordsFragment.kt
@@ -27,7 +27,7 @@ class DictionaryOfWordsFragment : BaseFragment() {
 
     private val dictionaryOfWordsPresenter by lazy {
         DictionaryOfWordsPresenter(
-            mainInteractor = MainInteractor(
+            mainInteractor = DictionaryOfWordsInteractor(
                 remoteRepository = DictionaryOfWordsRepository(
                     DictionaryApiHolder
                 ), localRepository = DictionaryOfWordsRepository(DictionaryApiHolder)
